Added classes

+++ Mined-Out/Engine/Models/PlayingField/FieldSize.cs
@@ -2,10 +2,10 @@
 {
     public struct FieldSize
     {
-        public double Width { get; set; }
-        public double Height { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
 
-        public FieldSize(double width, double height)
+        public FieldSize(int width, int height)
         {
             Width = width;
             Height = height;
